refactor

+++ podman-pilot/src/defaults.rs
@@ -26,7 +26,7 @@ use std::env;
 pub const CONTAINER_FLAKE_DIR: &str = "/usr/share/flakes";
 pub const CONTAINER_DIR: &str = "/var/lib/containers";
 pub const CONTAINER_CID_DIR: &str = "/var/lib/containers/storage/tmp/flakes";
-pub const GC_THRESHOLD: i32 = 20;
+pub const GC_THRESHOLD: usize = 20;
 pub const HOST_DEPENDENCIES: &str = "removed";
 
 pub fn debug(message: &str) {
